add styles for discount price in product card add types file in project
