+++ src/lib/utils/constants/products.ts
@@ -23,7 +23,7 @@ export const products = [
   },
   {
     id: '2',
-    slug: '1',
+    slug: '2',
     name: 'Корм для собак2',
     productCode: '1111',
     price: '800',
@@ -45,7 +45,7 @@ export const products = [
   },
   {
     id: '3',
-    slug: '1',
+    slug: '3',
     name: 'Корм для собак3',
     productCode: '1111',
     price: '800',
@@ -67,7 +67,7 @@ export const products = [
   },
   {
     id: '4',
-    slug: '1',
+    slug: '4',
     name: 'Корм для собак4',
     productCode: '1111',
     price: '800',
@@ -89,7 +89,7 @@ export const products = [
   },
   {
     id: '5',
-    slug: '1',
+    slug: '5',
     name: 'Корм для собак5',
     productCode: '1111',
     price: '800',
@@ -111,7 +111,7 @@ export const products = [
   },
   {
     id: '6',
-    slug: '1',
+    slug: '6',
     name: 'Корм для собак6',
     productCode: '1111',
     price: '800',
@@ -133,7 +133,7 @@ export const products = [
   },
   {
     id: '7',
-    slug: '1',
+    slug: '7',
     name: 'Корм для собак7',
     productCode: '1111',
     price: '800',
@@ -155,7 +155,7 @@ export const products = [
   },
   {
     id: '8',
-    slug: '1',
+    slug: '8',
     name: 'Корм для собак8',
     productCode: '1111',
     price: '800',
@@ -177,7 +177,7 @@ export const products = [
   },
   {
     id: '9',
-    slug: '1',
+    slug: '9',
     name: 'Корм для собак9',
     productCode: '1111',
     price: '800',
@@ -199,7 +199,7 @@ export const products = [
   },
   {
     id: '10',
-    slug: '1',
+    slug: '10',
     name: 'Корм для собак10',
     productCode: '1111',
     price: '800',
@@ -221,11 +221,11 @@ export const products = [
   },
   {
     id: '11',
-    slug: '1',
+    slug: '11',
     name: 'Корм для собак11',
     productCode: '1111',
     price: '800',
-    priceDiscount: '15',
+    priceDiscount: '680',
     categories: '',
     disciuntPercentage: '15',
     typesProduct: '',
@@ -243,11 +243,11 @@ export const products = [
   },
   {
     id: '12',
-    slug: '1',
+    slug: '12',
     name: 'Корм для собак12',
     productCode: '1111',
     price: '800',
-    priceDiscount: '10',
+    priceDiscount: '720',
     categories: '',
     disciuntPercentage: '10',
     typesProduct: '',
@@ -265,11 +265,11 @@ export const products = [
   },
   {
     id: '13',
-    slug: '1',
+    slug: '13',
     name: 'Корм для собак13',
     productCode: '1111',
     price: '800',
-    priceDiscount: '10',
+    priceDiscount: '720',
     categories: '',
     disciuntPercentage: '10',
     typesProduct: '',
@@ -287,11 +287,11 @@ export const products = [
   },
   {
     id: '14',
-    slug: '1',
+    slug: '14',
     name: 'Корм для собак14',
     productCode: '1111',
     price: '800',
-    priceDiscount: '10',
+    priceDiscount: '720',
     categories: '',
     disciuntPercentage: '10',
     typesProduct: '',
@@ -309,11 +309,11 @@ export const products = [
   },
   {
     id: '15',
-    slug: '1',
+    slug: '15',
     name: 'Корм для собак15',
     productCode: '1111',
     price: '800',
-    priceDiscount: '10',
+    priceDiscount: '720',
     categories: '',
     disciuntPercentage: '10',
     typesProduct: '',
@@ -331,11 +331,11 @@ export const products = [
   },
   {
     id: '16',
-    slug: '1',
+    slug: '16',
     name: 'Корм для собак16',
     productCode: '1111',
     price: '800',
-    priceDiscount: '10',
+    priceDiscount: '720',
     categories: '',
     disciuntPercentage: '10',
     typesProduct: '',
